+++ core/src/test/scala/tech/beshu/ror/integration/CurrentUserMetadataAccessControlTests.scala
@@ -191,7 +191,7 @@ class CurrentUserMetadataAccessControlTests
       |    bind_password: "password"                                 # skip for anonymous bind
       |    search_user_base_DN: "ou=People,dc=example,dc=com"
       |    search_groups_base_DN: "ou=Groups,dc=example,dc=com"
-      |    user_id_attribute: "cn"                                  # default "uid"
+      |    user_id_attribute: "uid"                                  # default "uid"
       |    unique_member_attribute: "uniqueMember"                   # default "uniqueMember"
       |    connection_pool_size: 10                                  # default 30
       |    connection_timeout_in_sec: 10                             # default 1
@@ -207,7 +207,7 @@ class CurrentUserMetadataAccessControlTests
       |    bind_password: "password"                                 # skip for anonymous bind
       |    search_user_base_DN: "ou=People,dc=example,dc=com"
       |    search_groups_base_DN: "ou=Groups,dc=example,dc=com"
-      |    user_id_attribute: "cn"                                  # default "uid"
+      |    user_id_attribute: "uid"                                  # default "uid"
       |    unique_member_attribute: "uniqueMember"                   # default "uniqueMember"
       |    connection_pool_size: 10                                  # default 30
       |    connection_timeout_in_sec: 10                             # default 1
