+++ core/src/test/scala/tech/beshu/ror/unit/acl/blocks/definitions/UnboundidLdapAuthenticationServiceTests.scala
@@ -16,28 +16,22 @@
  */
 package tech.beshu.ror.unit.acl.blocks.definitions
 
-import cats.data._
 import com.dimafeng.testcontainers.{Container, ForAllTestContainer, MultipleContainers}
-import com.unboundid.ldap.sdk.LDAPSearchException
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
-import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
-import monix.execution.exceptions.ExecutionRejectedException
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.{Assertion, BeforeAndAfterAll, BeforeAndAfterEach, Inside}
-import tech.beshu.ror.accesscontrol.blocks.definitions.CircuitBreakerConfig
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.LdapService.Name
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UnboundidLdapConnectionPoolProvider.LdapConnectionConfig._
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations.UserSearchFilterConfig.UserIdAttribute.CustomAttribute
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.implementations._
-import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.{CircuitBreakerLdapAuthenticationServiceDecorator, Dn, LdapAuthenticationService}
+import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.{Dn, LdapAuthenticationService}
 import tech.beshu.ror.accesscontrol.domain.{PlainTextSecret, User}
-import tech.beshu.ror.utils.ScalaOps.repeat
 import tech.beshu.ror.utils.SingletonLdapContainers
-import tech.beshu.ror.utils.containers.{LdapContainer, ToxiproxyContainer}
+import tech.beshu.ror.utils.containers.LdapContainer
 
 import java.time.Clock
 import scala.concurrent.duration._
@@ -51,14 +45,10 @@ class UnboundidLdapAuthenticationServiceTests
     with ForAllTestContainer
     with Inside {
 
-  private val ldap1ContainerWithToxiproxy = new ToxiproxyContainer(
-    SingletonLdapContainers.ldap1,
-    LdapContainer.defaults.ldap.port
-  )
-  private val ldapContainerToStop = LdapContainer.create("LDAP3", "test_example.ldif")
+  private val ldapContainer = LdapContainer.create("LDAP3", "test_example.ldif")
   private val ldapConnectionPoolProvider = new UnboundidLdapConnectionPoolProvider
 
-  override val container: Container = MultipleContainers(ldap1ContainerWithToxiproxy, ldapContainerToStop)
+  override val container: Container = MultipleContainers(ldapContainer)
 
   override protected def afterAll(): Unit = {
     super.afterAll()
@@ -66,105 +56,26 @@ class UnboundidLdapAuthenticationServiceTests
       .runSyncUnsafe()
   }
 
-  override protected def beforeEach(): Unit = {
-    super.beforeEach()
-    ldap1ContainerWithToxiproxy.enableNetwork()
-    ldap1ContainerWithToxiproxy.disableNetworkTimeout()
-  }
 
   "An LdapAuthenticationService" should {
     "have method to authenticate" which {
       "returns true" when {
-        "user exists in LDAP and its credentials are correct" ignore {
+        "user exists in LDAP and its credentials are correct" in {
           createSimpleAuthenticationService().assertSuccessfulAuthentication
-        }
-        "after connection timeout and retry" ignore {
-          val authenticationService = createSimpleAuthenticationService()
-          authenticationService.assertSuccessfulAuthentication
-          ldap1ContainerWithToxiproxy.enableNetworkTimeout()
-          authenticationService.assertFailedAuthentication[LdapUnexpectedResult]
-          ldap1ContainerWithToxiproxy.disableNetworkTimeout()
-          authenticationService.assertSuccessfulAuthentication
-        }
-        "after connection failure and retry" ignore {
-          val authenticationService = createSimpleAuthenticationService()
-          authenticationService.assertSuccessfulAuthentication
-          ldap1ContainerWithToxiproxy.disableNetwork()
-          authenticationService.assertFailedAuthentication[LDAPSearchException]
-          ldap1ContainerWithToxiproxy.enableNetwork()
-          authenticationService.assertSuccessfulAuthentication
         }
       }
       "returns false" when {
-        "user doesn't exist in LDAP" ignore {
+        "user doesn't exist in LDAP" in {
           createSimpleAuthenticationService()
             .authenticate(User.Id("unknown"), PlainTextSecret("user1"))
             .runSyncUnsafe() should be(false)
         }
-        "user has invalid credentials" ignore {
+        "user has invalid credentials" in {
           createSimpleAuthenticationService()
             .authenticate(User.Id("morgan"), PlainTextSecret("invalid_secret"))
             .runSyncUnsafe() should be(false)
         }
       }
-    }
-    "be able to work" when {
-      "Round robin HA method is configured" when {
-        "one of servers goes down" ignore {
-          def assertMorganCanAuthenticate(service: UnboundidLdapAuthenticationService) = {
-            service
-              .authenticate(User.Id("morgan"), PlainTextSecret("user1"))
-              .runSyncUnsafe() should be(true)
-          }
-          val service = createHaAuthenticationService()
-          (for {
-            _ <- repeat(maxRetries = 5, delay = 500 millis) {
-                Task(assertMorganCanAuthenticate(service))
-              }
-            _ <- Task(ldapContainerToStop.stop())
-            _ <- repeat(10, 500 millis) {
-              Task(assertMorganCanAuthenticate(service))
-            }
-          } yield ()).runSyncUnsafe()
-        }
-      }
-    }
-  }
-  "An CircuitBreaker decorated LdapAuthenticationService" should {
-    "close circuit breaker after 2 failed attempts" ignore {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      authenticationService.assertSuccessfulAuthentication
-      ldap1ContainerWithToxiproxy.disableNetwork()
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-      ldap1ContainerWithToxiproxy.enableNetwork()
-      authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-    }
-    "close circuit breaker after 2 failed attempts, but open it later" ignore {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      authenticationService.assertSuccessfulAuthentication
-      ldap1ContainerWithToxiproxy.disableNetwork()
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      ldap1ContainerWithToxiproxy.enableNetwork()
-      ldap1ContainerWithToxiproxy.enableNetworkTimeout()
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-      ldap1ContainerWithToxiproxy.disableNetworkTimeout()
-      Thread.sleep(550)
-      authenticationService.assertSuccessfulAuthentication
-      authenticationService.assertSuccessfulAuthentication
-    }
-    "close circuit breaker after 2 failed attempts and keep it closed because of network issues" ignore {
-      val authenticationService = createCircuitBreakerDecoratedSimpleAuthenticationService()
-      authenticationService.assertSuccessfulAuthentication
-      ldap1ContainerWithToxiproxy.disableNetwork()
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[ExecutionRejectedException]
-      Thread.sleep(550)
-      authenticationService.assertFailedAuthentication[LDAPSearchException]
-      authenticationService.assertFailedAuthentication[ExecutionRejectedException]
     }
   }
 
@@ -186,10 +97,14 @@ class UnboundidLdapAuthenticationServiceTests
     implicit val clock: Clock = Clock.systemUTC()
     UnboundidLdapAuthenticationService
       .create(
-        Name("my_ldap"),
+        Name("LDAP3"),
         ldapConnectionPoolProvider,
         LdapConnectionConfig(
-          ConnectionMethod.SingleServer(LdapHost.from(s"ldap://localhost:${ldap1ContainerWithToxiproxy.innerContainerMappedPort}").get),
+          ConnectionMethod.SingleServer(
+            LdapHost
+              .from(s"ldap://${SingletonLdapContainers.ldap1.ldapHost}:${SingletonLdapContainers.ldap1.ldapPort}")
+              .get
+          ),
           poolSize = 1,
           connectionTimeout = Refined.unsafeApply(5 seconds),
           requestTimeout = Refined.unsafeApply(5 seconds),
@@ -199,45 +114,6 @@ class UnboundidLdapAuthenticationServiceTests
             PlainTextSecret("password")
           ),
           ignoreLdapConnectivityProblems = false,
-        ),
-        UserSearchFilterConfig(Dn("ou=People,dc=example,dc=com"), CustomAttribute("uid"))
-      )
-      .runSyncUnsafe()
-      .getOrElse(throw new IllegalStateException("LDAP connection problem"))
-  }
-
-  private def createCircuitBreakerDecoratedSimpleAuthenticationService() = {
-    new CircuitBreakerLdapAuthenticationServiceDecorator(
-      createSimpleAuthenticationService(),
-      CircuitBreakerConfig(
-        maxFailures = Refined.unsafeApply(2),
-        resetDuration = Refined.unsafeApply(0.5 second))
-    )
-  }
-
-  private def createHaAuthenticationService() = {
-    implicit val clock: Clock = Clock.systemUTC()
-    UnboundidLdapAuthenticationService
-      .create(
-        Name("my_ldap"),
-        ldapConnectionPoolProvider,
-        LdapConnectionConfig(
-          ConnectionMethod.SeveralServers(
-            NonEmptyList.of(
-              LdapHost.from(s"ldap://${SingletonLdapContainers.ldap1.ldapHost}:${SingletonLdapContainers.ldap1.ldapPort}").get,
-              LdapHost.from(s"ldap://${ldapContainerToStop.ldapHost}:${ldapContainerToStop.ldapPort}").get,
-            ),
-            HaMethod.RoundRobin
-          ),
-          poolSize = 1,
-          connectionTimeout = Refined.unsafeApply(5 seconds),
-          requestTimeout = Refined.unsafeApply(5 seconds),
-          trustAllCerts = false,
-          BindRequestUser.CustomUser(
-            Dn("cn=admin,dc=example,dc=com"),
-            PlainTextSecret("password")
-          ),
-          ignoreLdapConnectivityProblems = false
         ),
         UserSearchFilterConfig(Dn("ou=People,dc=example,dc=com"), CustomAttribute("uid"))
       )
